+++ relay/kusama/constants/src/lib.rs
@@ -23,7 +23,7 @@ pub mod currency {
 	use polkadot_primitives::Balance;
 
 	/// The existential deposit.
-	pub const EXISTENTIAL_DEPOSIT: Balance = 1 * CENTS;
+	pub const EXISTENTIAL_DEPOSIT: Balance = CENTS;
 
 	pub const UNITS: Balance = 1_000_000_000_000;
 	pub const QUID: Balance = UNITS / 30;
@@ -42,7 +42,7 @@ pub mod time {
 	use polkadot_runtime_common::prod_or_fast;
 	pub const MILLISECS_PER_BLOCK: Moment = 6000;
 	pub const SLOT_DURATION: Moment = MILLISECS_PER_BLOCK;
-	pub const EPOCH_DURATION_IN_SLOTS: BlockNumber = prod_or_fast!(1 * HOURS, 1 * MINUTES);
+	pub const EPOCH_DURATION_IN_SLOTS: BlockNumber = prod_or_fast!(HOURS, MINUTES);
 
 	// These time units are defined in number of blocks.
 	pub const MINUTES: BlockNumber = 60_000 / (MILLISECS_PER_BLOCK as BlockNumber);