+++ system-parachains/bridge-hubs/bridge-hub-polkadot/src/weights/xcm/mod.rs
@@ -19,7 +19,6 @@ mod pallet_xcm_benchmarks_generic;
 
 use crate::{xcm_config::MaxAssetsIntoHolding, Runtime};
 use codec::Encode;
-use frame_support::weights::Weight;
 use pallet_xcm_benchmarks_fungible::WeightInfo as XcmFungibleWeight;
 use pallet_xcm_benchmarks_generic::WeightInfo as XcmGeneric;
 use sp_std::prelude::*;